+++ examples/inference/basic/basic_cosmos2_5_i2w.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 from fastvideo import VideoGenerator
-from fastvideo.configs.sample import SamplingParam
+from fastvideo.api.sampling_param import SamplingParam
 
 
 def main():
