+++ examples/inference/basic/basic_cosmos2_5_t2w.py
@@ -1,5 +1,5 @@
 from fastvideo import VideoGenerator
-from fastvideo.configs.sample import SamplingParam
+from fastvideo.api.sampling_param import SamplingParam
 
 
 def main():
