+++ examples/inference/lora/wan_lora_inference_from_ckpt.py
@@ -2,7 +2,7 @@
 Inference using a LoRA checkpoint from FastVideo trainer.
 """
 from fastvideo import VideoGenerator
-from fastvideo.configs.sample import SamplingParam
+from fastvideo.api.sampling_param import SamplingParam
 
 OUTPUT_PATH = "./lora_out"
 def main():
