+++ fastvideo/__init__.py
@@ -1,5 +1,5 @@
 from fastvideo.configs.pipelines import PipelineConfig
-from fastvideo.configs.sample import SamplingParam
+from fastvideo.api.sampling_param import SamplingParam
 from fastvideo.entrypoints.video_generator import VideoGenerator
 from fastvideo.version import __version__
