+++ fastvideo/entrypoints/streaming_generator.py
@@ -8,7 +8,7 @@ import torch
 import torchvision
 from einops import rearrange
 
-from fastvideo.configs.sample import SamplingParam
+from fastvideo.api.sampling_param import SamplingParam
 from fastvideo.entrypoints.video_generator import VideoGenerator
 from fastvideo.fastvideo_args import FastVideoArgs
 from fastvideo.logger import init_logger
