+++ fastvideo/pipelines/preprocess/preprocess_pipeline_ode_trajectory.py
@@ -18,7 +18,7 @@ from torch.utils.data import DataLoader
 from torchdata.stateful_dataloader import StatefulDataLoader
 from tqdm import tqdm
 
-from fastvideo.configs.sample import SamplingParam
+from fastvideo.api.sampling_param import SamplingParam
 from fastvideo.dataset import gettextdataset
 from fastvideo.dataset.dataloader.parquet_io import (ParquetDatasetWriter, records_to_table)
 from fastvideo.dataset.dataloader.record_schema import (ode_text_only_record_creator)
