+++ fastvideo/training/wan_i2v_distillation_pipeline.py
@@ -6,7 +6,7 @@ from typing import Any
 import torch
 from einops import rearrange
 
-from fastvideo.configs.sample import SamplingParam
+from fastvideo.api.sampling_param import SamplingParam
 from fastvideo.dataset.dataloader.schema import pyarrow_schema_i2v
 from fastvideo.distributed import get_local_torch_device
 from fastvideo.fastvideo_args import FastVideoArgs, TrainingArgs
